refactor: Move template selector to a separate file
